Modularize Game Loop
